+++ packages/qwik/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the library will be documented in this file.
 
-## vX.X.X (Month DD, YYYY)
+## v0.4.0 (March 25, 2023)
 
 - Extend parameter of `zodField$` and `zodForm$` to allow passing a function (issue #38)
 - Add `data` property to `FormResponse` type (issue #37)